Read animated GIFs via a new GifFrameReader class.
As per previous discussions, I am also changing the MultipleFrameReader
interface so that Get{Image,Frame}Spec() assign a copy rather than a
pointer. Some of the other changes made it obvious this was the time to do it.

This CL depends on the changes in
https://code.google.com/p/page-speed/source/detail?r=2578
(checked-in, thank you bmcquade@!) and in
http://page-speed-codereview.appspot.com/1170001/ to support, in
ModPageSpeed-land, the giflib encryption target needed by the tests.

+++ src/pagespeed/kernel/image/image_util.cc
@@ -64,7 +64,7 @@ const char* GetPixelFormatString(PixelFormat pixel_format) {
   return kInvalidPixelFormat;
 }
 
-int GetBytesPerPixel(PixelFormat pixel_format) {
+size_t GetBytesPerPixel(PixelFormat pixel_format) {
   switch (pixel_format) {
     case UNSUPPORTED: return 0;
     case RGB_888:     return 3;
